# doc/swagger/swagger-latest.yaml
swagger: '2.0'
info:
  version: 0.0.5
  title: Notch Health D2C 
  description: >
    ## Overview

    This specification defines the necessary resources needed to access a patient's information and the patient's test results.
    
    
    A customer orders a test kit through a website, when the test kit order is received by the customer, they register each kit to a person who will take the test. When registering the kit, they will enter a test kit ID along with their name, date of birth and other personal information. 
    
    Upon submission of the kit ID, a lab order is created. Once the customer has administered the test, they mail back the kit to the lab.
    
    
    The customer can track the kit to the lab, check the status or the results, update any personal information related to the lab order.
    
    
    Eight services exist to that act as a broker of information between the lab system and the website.
    
    1) A service to create a kit order request
    
    2) A service to request the delivery status of a kit
    
    3) A service to update the kit order with kit IDs
    
    4) A service to create a lab order
    
    5) A service to get all lab orders for a customer
    
    6) A service to get a specific lab order (to check the status)
    
    7) A service to update a lab order
    
    8) A service to delete a lab order
    
    9) A service to check the existence of a lab order

    10) A service to get the customer's Shopify orders

...

...

...

...

...


...


...


...


...

...

...
host: "notchapi-service-ovvelzitca-uw.a.run.app"
basePath: / 
schemes:
...
produces:
...
consumes:
...
paths:
    /test-kit-orders:
      post:
        tags:
          - test-kit-order
        parameters:
          - name: testKit
            in: body
            description: "User events captured"
            required: true
            schema:
              type: object
              items:
                $ref: '#/definitions/test-kit'
        summary: "Create test kit order"
        operationId: testKitOrderPost
        responses:
          '201':
            description: test kit created
    /test-kit-orders/{id}:
      get:
        tags:
          - test-kit-order
        parameters:
          - in: path
            name: id
            type: string
            required: true
            description: "ID of test kit"
        summary: "Get the tracking information based on test kit id"
        operationId: testKitOrderGetByID
        responses:
          '200':
            description: "Test kit object"
      post:
        tags:
          - test-kit-order
        parameters:
          - name: id
            in: path
            description: "ID of test kit"
            required: true
            type: string
          - name: kit
            in: body
            description: "Kit ID/s to be associated with kit order"
            required: true
            schema:
              type: object
              items:
                $ref: '#/definitions/test-kit'
        summary: "Update the test kit order with test kit id/s"
        operationId: testKitOrderUpdateByID
        responses:
          '200':
            description: "Test kit object"
...
    /orders/{id}:
      get:
        tags:
          - order
        parameters:
          - name: id
            in: path
            description: "ID of the shopify order"
            required: true
            type: string
        summary: "Get a specified shopify order based on order ID"
        operationId: ShopifyOrderGetByID
        responses:
          '200':
            description: "response object"
    /orders:
      post:
        tags:
          - order
        parameters:
          - name: email
            in: body
            description: "email address associated to the shopify order"
            required: true
            schema:
              type: string
        summary: "Get a specified shopify order based on order email"
        operationId: ShopifyOrderGetByEmail
        responses:
          '200':
            description: "response object"
...
definitions:
  test-kit:
    type: object
    properties:
      id:
        type: string
      order-id:
        type: string
      customer-id:
        type: string
      pract-id:
        type: string
      order-date:
        type: string
      order-taker:
        type: string
      requester:
        type: string
      process-type:
        type: string
      loc-pos:
        type: string
      clinic-name:
        type: string
      address1:
        type: string
      address2:
        type: string
      city:
        type: string
      state:
        type: string
      other-state:
        type: string
      zip:
        type: string
      country:
        type: string
      req-loc:
        type: string
      shipping-method:
        type: string
      tracking-num:
        type: string
      status:
        type: string
      comment:
        type: string
      kit-id:
        type: string
...